refactor: change interface file names
