+++ src/shared/libs/file-writer/index.ts
@@ -1,2 +1,2 @@
-export * from './file-writer-interface.js';
+export * from './file-writer.interface.js';
 export * from './tsv-file-writer.js';
