+++ src/shared/libs/file-writer/tsv-file-writer.ts
@@ -1,5 +1,5 @@
 import { WriteStream, createWriteStream } from 'node:fs';
-import { FileWriter } from './file-writer-interface.js';
+import { FileWriter } from './file-writer.interface.js';
 
 export class TSVFileWriter implements FileWriter {
   private stream: WriteStream;
