+++ src/shared/libs/offer-generator/index.ts
@@ -1,1 +1,1 @@
-export * from './offer-generator-interface.js';
+export * from './offer-generator.interface.js';
