+++ src/shared/libs/offer-generator/offer-generator.interface.ts
